requirements.txt
scipy==1.9.3
absl-py==1.2.0
accelerate
aiofiles==23.2.1
aiohttp==3.9.4
aiohttp-cors==0.7.0
aiosignal==1.3.1
airportsdata==20241001
altair==5.3.0
annotated-types==0.6.0
anthropic==0.37.1
antlr4-python3-runtime==4.11.1
anyio==4.3.0
appdirs==1.4.4
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
autoawq
autoawq_kernels
av==13.1.0
Babel==2.15.0
backcall==0.2.0
backoff==2.2.1
bcrypt==4.1.3
beautifulsoup4==4.12.3
bitsandbytes==0.43.1
blake3==1.0.1
bleach==6.1.0
build==1.2.1
cachetools==5.3.3
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.5.0
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloudpickle==3.1.0
coloredlogs==15.0.1
colorful==0.5.6
colormath==3.0.0
comm==0.2.2
compressed-tensors==0.9.1
contourpy==1.2.1
cupy-cuda12x==13.4.0
dataclasses-json==0.6.6
datasets==2.21.0
debugpy==1.8.1
decorator==5.1.1
deepspeed==0.15.3
defusedxml==0.7.1
Deprecated==1.2.14
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
Django==5.0.6
django-emoji==2.2.2
dnspython==2.7.0
docker-pycreds==0.4.0
docopt==0.6.2
docstring_parser==0.16
einops==0.7.0
email_validator==2.2.0
exceptiongroup==1.2.0
executing==2.0.1
Farama-Notifications==0.0.4
fastapi==0.115.7
fastapi-cli==0.0.7
fastchat==0.1.0
fastjsonschema==2.19.1
fastrlock==0.8.3
ffmpy==0.3.2
filelock==3.16.1
fiona==1.10.1
fire==0.6.0
flashinfer
flatbuffers==24.3.25
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2024.2.0
generativeai==0.0.1
geo==0.1.1
geopandas==0.11.1
gguf==0.10.0
git-lfs==1.6
gitdb==4.0.11
GitPython==3.1.43
google-ai-generativelanguage==0.6.10
google-api-core==2.19.0
google-api-python-client==2.130.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.63.0
gradio==5.12.0
gradio_client==1.5.4
greenlet==3.0.3
grpcio==1.64.0
grpcio-status==1.62.2
gym==0.26.2
gym-notices==0.0.8
gymnasium==1.0.0
h11==0.14.0
hjson==3.1.0
html2text==2024.2.26
httpcore==1.0.5
httplib2==0.22.0
httptools==0.6.1
httpx==0.27.0
huggingface-hub==0.27.1
humanfriendly==10.0
ijson==3.3.0
importlib-metadata==7.0.0
importlib_resources==6.4.0
iniconfig==2.0.0
interegular==0.3.3
ipykernel==6.29.4
ipython==8.12.3
ipywidgets==8.1.2
isoduration==20.11.0
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.5
jiter==0.5.0
joblib==1.4.0
json5==0.9.25
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.10.0
jupyter-lsp==2.2.5
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyter_server==2.14.0
jupyter_server_terminals==0.5.3
jupyterlab==4.2.1
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.2
jupyterlab_widgets==3.0.10
kiwisolver==1.4.5
kubernetes==29.0.0
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.1
langchain-text-splitters==0.2.0
langsmith==0.1.63
lark==1.2.2
libpysal==4.6.2
linkify-it-py==2.0.3
llvmlite==0.43.0
lm-format-enforcer==0.10.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
mistral_common==1.5.1
mistune==3.0.2
mmh3==4.1.0
momepy==0.5.3
monotonic==1.6
msgpack==1.1.0
msgspec==0.18.6
multidict==6.0.5
multipledispatch==1.0.0
multiprocess==0.70.16
munkres==1.1.4
mypy-extensions==1.0.0
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==2.8.5
ninja==1.11.1.1
nltk==3.8.1
notebook==7.2.0
notebook_shim==0.2.4
numba==0.60.0
numpy==1.25.2
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
oauthlib==3.2.2
onnxruntime==1.18.0
openai==1.52.2
opencensus==0.11.4
opencensus-context==0.1.3
opencv-python-headless==4.10.0.84
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.10.0
osmnx==1.1.2
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==23.2
pandas==1.4.3
pandocfilters==1.5.1
parso==0.8.4
partial-json-parser==0.2.1.1.post4
Pebble==5.0.7
peft==0.11.1
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.3.0
platformdirs==4.2.2
pluggy==1.5.0
posthog==3.5.0
prometheus-fastapi-instrumentator==7.0.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
py-spy==0.4.0
pyairports==2.1.1
pyarrow==15.0.2
pyarrow-hotfix==0.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pybind11==2.13.6
pycountry==24.6.1
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pygeos==0.14
Pygments==2.17.2
pymannkendall==1.4.3
pyogrio==0.10.0
pyparsing==3.1.2
PyPika==0.48.9
pyproj==3.7.0
pyproject_hooks==1.1.0
pyshp==2.3.1
pytest==8.3.4
python-dotenv==1.0.1
python-json-logger==2.0.7
python-multipart==0.0.20
PyYAML==6.0.1
pyzmq==26.0.3
qtconsole==5.5.2
QtPy==2.4.1
ray==2.40.0
referencing==0.34.0
regex==2023.12.25
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.1
rich-toolkit==0.13.2
rouge-chinese==1.0.3
rpds-py==0.18.0
rsa==4.9
rtree==1.4.0
ruff==0.3.7
safehttpx==0.1.6
safetensors==0.5.2
scikit-learn==1.5.0
semantic-version==2.10.0
Send2Trash==1.8.3
sentence-transformers==2.7.0
sentencepiece==0.2.0
sentry-sdk==2.19.2
setproctitle==1.3.4
Shapely==1.8.2
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smart-open==7.1.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
sse-starlette==2.1.0
stable_baselines3==2.4.0
stack-data==0.6.3
starlette==0.45.2
sumolib==1.21.0
sympy==1.13.1
tempdir==0.7.1
tenacity==8.3.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
termcolor==2.4.0
terminado==0.18.1
textual==1.0.0
threadpoolctl==3.5.0
tiktoken==0.7.0
timeout-decorator==0.5.0
tinycss2==1.3.0
tokenizers==0.21.0
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
tornado==6.4
tqdm==4.66.6
traci==1.21.0
traitlets==5.14.3
transformers==4.48.2
tree-sitter==0.23.2
tree-sitter-python==0.23.2
triton==3.1.0
trl==0.9.6
typer==0.12.3
types-python-dateutil==2.9.0.20240316
typing-inspect==0.9.0
tyro==0.8.3
tzdata==2024.1
uc-micro-py==1.0.3
uri-template==1.3.0
uritemplate==4.1.1
urllib3==2.2.1
uvicorn==0.29.0
uvloop==0.19.0
virtualenv==20.28.1
vllm==0.8.5
wandb==0.19.0
watchfiles==0.21.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wget==3.2
widgetsnbextension==4.0.10
word2number==1.1
wrapt==1.16.0
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
zipp==3.19.0
zstandard==0.23.0
geos